+++ meetupselector/api/routes/users.py
@@ -3,7 +3,8 @@ from http import HTTPStatus
 from ninja import Router
 from ninja.security import django_auth
 
-from meetupselector.api.schemas.users import LoginSchema, SignInSchema
+from meetupselector.api.schemas.users import LoginSchema
+from meetupselector.user.schemas import SignInSchema
 from meetupselector.user.services import UserService
 
 router = Router(auth=django_auth)
@@ -23,4 +24,4 @@ def login(request, credentials: LoginSchema, auth=None):
 
 @router.post("/users", response={HTTPStatus.CREATED: None}, url_name="create_user", auth=None)
 def create_user(_, credentials: SignInSchema):
-    return HTTPStatus.CREATED, UserService.create(**credentials.dict())
+    return HTTPStatus.CREATED, UserService.create(credentials)
